requirements.txt
prometheus-client==0.17.0
cs50
flask
flask-session
openai
python-dotenv
requests
werkzeug